refactor: admin Pages type 수정
